requirements.txt
openai
tqdm
pandas
python-dotenv
matplotlib
seaborn